Dravon/requirements.txt
discord.py>=2.6.3
wavelink>=3.4.1
motor>=3.3.2
aiohttp>=3.12.15
async-timeout>=5.0.1